fix(security): API 키 노출, 봇 검증 우회, 권한 검증 무력화 세 결함 보강
1) 복호화된 AI API 키 조회 함수를 RSC action 으로 자동 노출되지 않는 server-only 모듈로 분리합니다.

   기존에는 해당 함수가 서버 액션 파일 안에 export 되어 있어 Next.js 가 자동으로 RSC action 엔드포인트로 등록했고, 클라이언트에서 임의 프로젝트 식별자를 넘겨 호출하면 평문 API 키를 받아갈 수 있는 노출 경로가 열려 있었습니다. 서버 내부 호출은 라우트 핸들러 한 곳뿐이라 import 경로만 옮기면 됩니다.

2) 프로젝트 생성과 프로젝트 접근 검증 두 진입점의 봇 검증을 필수화합니다.

   기존에는 봇 검증 토큰이 있을 때만 검증을 수행했고, 토큰을 보내지 않으면 검증이 통째로 우회됐습니다. 토큰이 없으면 진입 자체를 차단하도록 가드를 앞에 둡니다.

3) 프로젝트 접근 권한 검증 함수를 활성화합니다.

   기존에는 검증 함수가 임시로 항상 통과를 반환하고 있어, 모든 mutation 서버 액션의 권한 검증이 사실상 비어 있었습니다. 다행히 데이터베이스 정책이 외부에서의 직접 접근을 막아 최악의 시나리오는 차단된 상태였지만, 서버 액션 경유 우회는 가능한 상태였습니다. 쿠키에 저장된 접근 토큰을 검증해 프로젝트 식별자가 일치할 때만 통과시키는 정상 로직을 복원합니다.
