+++ apps/web/src/access/lib/require-access.ts
@@ -4,9 +4,8 @@
  * 쿠키에 저장된 접근 토큰을 검증하여 해당 프로젝트에 대한 접근 권한이 있는지 확인.
  * 모든 프로젝트 데이터 변경(mutation) 서버 액션에서 호출해야 함.
  */
-
-import { getAllAccessTokenCookies } from './cookies';
 import { verifyProjectAccessToken } from './access-token';
+import { getAllAccessTokenCookies } from './cookies';
 
 /**
  * 프로젝트 ID 기반 접근 권한 확인
@@ -17,11 +16,7 @@ import { verifyProjectAccessToken } from './access-token';
  * @param projectId - 접근 대상 프로젝트 ID
  * @returns 접근 가능 여부
  */
-export async function requireProjectAccess(_projectId: string): Promise<boolean> {
-  // TODO: 로그인/인증 구현 후 아래 토큰 검증 로직 활성화
-  return true;
-
-  /*
+export async function requireProjectAccess(projectId: string): Promise<boolean> {
   try {
     const tokenMap = await getAllAccessTokenCookies();
 
@@ -36,5 +31,4 @@ export async function requireProjectAccess(_projectId: string): Promise<boolean>
   } catch {
     return false;
   }
-  */
 }
